refactor: use isItemAllowed
