+++ packages/widget/src/utils/item.ts
@@ -1,8 +1,33 @@
-import type { AllowDeny } from '../types/widget.js'
+import type { BaseToken } from '@lifi/sdk'
+import type { FormType } from '../stores/form/types.js'
+import type { AllowDeny, WidgetTokens } from '../types/widget.js'
 
-export const isItemAllowed = <T>(itemId: T, items?: AllowDeny<T>) => {
+type IncludesFn<T> = (list: T[], item: T) => boolean
+
+export const isItemAllowed = <T>(
+  item: T,
+  items?: AllowDeny<T>,
+  includes: IncludesFn<T> = (list, val) => list.includes(val)
+): boolean => {
   if (items?.allow?.length) {
-    return items.allow.includes(itemId)
+    return includes(items.allow, item)
   }
-  return !items?.deny?.includes(itemId)
+
+  return !includes(items?.deny ?? [], item)
+}
+
+const tokenIncludes = (list: BaseToken[], item: BaseToken) =>
+  list.some((t) => t.address === item.address && t.chainId === item.chainId)
+
+export const isTokenAllowed = (
+  token: BaseToken,
+  configTokens: WidgetTokens | undefined,
+  formType: FormType | undefined
+) => {
+  return (
+    isItemAllowed(token, configTokens, tokenIncludes) &&
+    (formType
+      ? isItemAllowed(token, configTokens?.[formType], tokenIncludes)
+      : true)
+  )
 }
